chore: move all cli functions into a single location
