+++ pyproject.toml
@@ -312,86 +312,86 @@ Homepage = "https://github.com/NVIDIA-NeMo/Gym"
 
 [project.scripts]
 # Run/test scripts for servers.
-nemo_gym_run = "nemo_gym.cli:run"
-ng_run = "nemo_gym.cli:run"
-nemo_gym_test = "nemo_gym.cli:test"
-ng_test = "nemo_gym.cli:test"
-nemo_gym_test_all = "nemo_gym.cli:test_all"
-ng_test_all = "nemo_gym.cli:test_all"
+nemo_gym_run = "nemo_gym.cli.env:run"
+ng_run = "nemo_gym.cli.env:run"
+nemo_gym_test = "nemo_gym.cli.env:test"
+ng_test = "nemo_gym.cli.env:test"
+nemo_gym_test_all = "nemo_gym.cli.env:test_all"
+ng_test_all = "nemo_gym.cli.env:test_all"
 
 # Development convenience aliases.
-nemo_gym_dev_test = "nemo_gym.cli:dev_test"
-ng_dev_test = "nemo_gym.cli:dev_test"
-nemo_gym_init_resources_server = "nemo_gym.cli:init_resources_server"
-ng_init_resources_server = "nemo_gym.cli:init_resources_server"
+nemo_gym_dev_test = "nemo_gym.cli.dev:dev_test"
+ng_dev_test = "nemo_gym.cli.dev:dev_test"
+nemo_gym_init_resources_server = "nemo_gym.cli.env:init_resources_server"
+ng_init_resources_server = "nemo_gym.cli.env:init_resources_server"
 
 # Benchmarks
-nemo_gym_list_benchmarks = "nemo_gym.benchmarks:list_benchmarks"
-ng_list_benchmarks = "nemo_gym.benchmarks:list_benchmarks"
-nemo_gym_prepare_benchmark = "nemo_gym.benchmarks:prepare_benchmark"
-ng_prepare_benchmark = "nemo_gym.benchmarks:prepare_benchmark"
+nemo_gym_list_benchmarks = "nemo_gym.cli.eval:list_benchmarks"
+ng_list_benchmarks = "nemo_gym.cli.eval:list_benchmarks"
+nemo_gym_prepare_benchmark = "nemo_gym.cli.eval:prepare_benchmark"
+ng_prepare_benchmark = "nemo_gym.cli.eval:prepare_benchmark"
 
 # Rollout collection
-nemo_gym_collect_rollouts = "nemo_gym.rollout_collection:collect_rollouts"
-ng_collect_rollouts = "nemo_gym.rollout_collection:collect_rollouts"
-nemo_gym_e2e_collect_rollouts = "nemo_gym.cli:e2e_rollout_collection"
-ng_e2e_collect_rollouts = "nemo_gym.cli:e2e_rollout_collection"
-nemo_gym_aggregate_rollouts = "nemo_gym.rollout_collection:aggregate_rollouts"
-ng_aggregate_rollouts = "nemo_gym.rollout_collection:aggregate_rollouts"
+nemo_gym_collect_rollouts = "nemo_gym.cli.eval:collect_rollouts"
+ng_collect_rollouts = "nemo_gym.cli.eval:collect_rollouts"
+nemo_gym_e2e_collect_rollouts = "nemo_gym.cli.eval:e2e_rollout_collection"
+ng_e2e_collect_rollouts = "nemo_gym.cli.eval:e2e_rollout_collection"
+nemo_gym_aggregate_rollouts = "nemo_gym.cli.eval:aggregate_rollouts"
+ng_aggregate_rollouts = "nemo_gym.cli.eval:aggregate_rollouts"
 
 # Prompt materialization
-nemo_gym_materialize_prompts = "nemo_gym.prompt:materialize_prompts_cli"
-ng_materialize_prompts = "nemo_gym.prompt:materialize_prompts_cli"
+nemo_gym_materialize_prompts = "nemo_gym.cli.dataset:materialize_prompts_cli"
+ng_materialize_prompts = "nemo_gym.cli.dataset:materialize_prompts_cli"
 
 # Reward profiling
-nemo_gym_reward_profile = "nemo_gym.reward_profile:reward_profile"
-ng_reward_profile = "nemo_gym.reward_profile:reward_profile"
+nemo_gym_reward_profile = "nemo_gym.cli.eval:reward_profile"
+ng_reward_profile = "nemo_gym.cli.eval:reward_profile"
 
 # Dataset management
-nemo_gym_upload_dataset_to_gitlab = "nemo_gym.gitlab_utils:upload_jsonl_dataset_cli"
-ng_upload_dataset_to_gitlab = "nemo_gym.gitlab_utils:upload_jsonl_dataset_cli"
-nemo_gym_download_dataset_from_gitlab = "nemo_gym.gitlab_utils:download_jsonl_dataset_cli"
-ng_download_dataset_from_gitlab = "nemo_gym.gitlab_utils:download_jsonl_dataset_cli"
-nemo_gym_prepare_data = "nemo_gym.train_data_utils:prepare_data"
-ng_prepare_data = "nemo_gym.train_data_utils:prepare_data"
+nemo_gym_upload_dataset_to_gitlab = "nemo_gym.cli.dataset:upload_jsonl_dataset_cli"
+ng_upload_dataset_to_gitlab = "nemo_gym.cli.dataset:upload_jsonl_dataset_cli"
+nemo_gym_download_dataset_from_gitlab = "nemo_gym.cli.dataset:download_jsonl_dataset_cli"
+ng_download_dataset_from_gitlab = "nemo_gym.cli.dataset:download_jsonl_dataset_cli"
+nemo_gym_prepare_data = "nemo_gym.cli.dataset:prepare_data"
+ng_prepare_data = "nemo_gym.cli.dataset:prepare_data"
 
 # HF dataset upload and download
-nemo_gym_upload_dataset_to_hf = "nemo_gym.dataset_orchestrator:upload_jsonl_dataset_to_hf_cli"
-ng_upload_dataset_to_hf = "nemo_gym.dataset_orchestrator:upload_jsonl_dataset_to_hf_cli"
-nemo_gym_download_dataset_from_hf = "nemo_gym.dataset_orchestrator:download_jsonl_dataset_from_hf_cli"
-ng_download_dataset_from_hf = "nemo_gym.dataset_orchestrator:download_jsonl_dataset_from_hf_cli"
+nemo_gym_upload_dataset_to_hf = "nemo_gym.cli.dataset:upload_jsonl_dataset_to_hf_cli"
+ng_upload_dataset_to_hf = "nemo_gym.cli.dataset:upload_jsonl_dataset_to_hf_cli"
+nemo_gym_download_dataset_from_hf = "nemo_gym.cli.dataset:download_jsonl_dataset_from_hf_cli"
+ng_download_dataset_from_hf = "nemo_gym.cli.dataset:download_jsonl_dataset_from_hf_cli"
 
 # Gitlab -> HF (upload then delete)
-nemo_gym_gitlab_to_hf_dataset = "nemo_gym.dataset_orchestrator:upload_jsonl_dataset_to_hf_and_delete_gitlab_cli"
-ng_gitlab_to_hf_dataset = "nemo_gym.dataset_orchestrator:upload_jsonl_dataset_to_hf_and_delete_gitlab_cli"
+nemo_gym_gitlab_to_hf_dataset = "nemo_gym.cli.dataset:upload_jsonl_dataset_to_hf_and_delete_gitlab_cli"
+ng_gitlab_to_hf_dataset = "nemo_gym.cli.dataset:upload_jsonl_dataset_to_hf_and_delete_gitlab_cli"
 
 # Manually delete from Gitlab
-nemo_gym_delete_dataset_from_gitlab = "nemo_gym.dataset_orchestrator:delete_jsonl_dataset_from_gitlab_cli"
-ng_delete_dataset_from_gitlab = "nemo_gym.dataset_orchestrator:delete_jsonl_dataset_from_gitlab_cli"
+nemo_gym_delete_dataset_from_gitlab = "nemo_gym.cli.dataset:delete_jsonl_dataset_from_gitlab_cli"
+ng_delete_dataset_from_gitlab = "nemo_gym.cli.dataset:delete_jsonl_dataset_from_gitlab_cli"
 
 # Configuration utils
-nemo_gym_dump_config = "nemo_gym.cli:dump_config"
-ng_dump_config = "nemo_gym.cli:dump_config"
+nemo_gym_dump_config = "nemo_gym.cli.env:dump_config"
+ng_dump_config = "nemo_gym.cli.env:dump_config"
 
 # Display help
-nemo_gym_help = "nemo_gym.cli:display_help"
-ng_help = "nemo_gym.cli:display_help"
+nemo_gym_help = "nemo_gym.cli.general:display_help_legacy"
+ng_help = "nemo_gym.cli.general:display_help_legacy"
 
 # Server status
-nemo_gym_status = "nemo_gym.cli:status"
-ng_status = "nemo_gym.cli:status"
+nemo_gym_status = "nemo_gym.cli.env:status"
+ng_status = "nemo_gym.cli.env:status"
 
 # Environment-specific uv pip list
-nemo_gym_pip_list = "nemo_gym.cli:pip_list"
-ng_pip_list = "nemo_gym.cli:pip_list"
+nemo_gym_pip_list = "nemo_gym.cli.env:pip_list"
+ng_pip_list = "nemo_gym.cli.env:pip_list"
 
 # Display version
-nemo_gym_version = "nemo_gym.cli:version"
-ng_version = "nemo_gym.cli:version"
+nemo_gym_version = "nemo_gym.cli.general:version"
+ng_version = "nemo_gym.cli.general:version"
 
 # Re-install Gym and dependencies
-nemo_gym_reinstall = "nemo_gym.cli:reinstall"
-ng_reinstall = "nemo_gym.cli:reinstall"
+nemo_gym_reinstall = "nemo_gym.cli.general:reinstall"
+ng_reinstall = "nemo_gym.cli.general:reinstall"
 
 [tool.setuptools.packages.find]
 where = ["."]
